ext method for sha1value
Former-commit-id: 15f78d53e21cdf65cdccd910373c1548120dbd08 [formerly e79cb2078f15ecc1d6c9af58ed4cb8158172d10e]
Former-commit-id: 321fc6e9fcbf1779bb44229f6e3085a6fa92dc12

+++ ExClient/Category.cs
@@ -4,7 +4,7 @@ using System.Text;
 namespace ExClient
 {
     [Flags]
-    public enum Category : uint
+    public enum Category : ushort
     {
         Unspecified = 0,
         Doujinshi = 0x01,
